test(moe): cover gate load across checkpoint modes
